Split certificate module out of crate root
